Push initial implementation of multi-value timeline based on uPlot
This utilizes some of Grafana's internal APIs which may (and will)
change in the future. It's not stable nor documented properly, so
this commit is only for future reference and it's going to be
reimplemented using other plot library. Further implementation
over existing internal Grafana's APIs would mean
reverse-engineering everything from built-in panel plugins and
creating a lot of dirty hacks to overcome the limits. I still
wanted to try it, because other libraries and not great to work
with anyway, but it's not reasonable anymore.

Reimplementation in other library comes with drawbacks: Grafana
theme is not going to be fully respected and we will have to
create our own bindings to various events (in React, etc.).